livekit/token_requirement.txt
# Token Server Requirements
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
livekit>=0.11.0
pydantic>=2.0.0
